fix: display only metadata
