+++ packages/cardano-metamask-snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/nufi-official/nufi-snap"
   },
   "source": {
-    "shasum": "MP8gYtCTLabhKQN+OFOvkiG+hj3IkYStKvvIVQwDjM4=",
+    "shasum": "ZZlBZ4csi8GJRjG3kdL0KjJqYl5vT45aRLOMwqBxW70=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
