+++ packages/cardano-metamask-snap/src/api/cardano__signTransaction/index.test.ts
@@ -148,8 +148,8 @@ const fixtures = {
     ownAddressParams: [],
     stringifiedUi: `{"type":"confirmation","content":{"type":"Box","props":{"children":[{"type":"Heading","props":{"children":"Sign transaction"},"key":null},{"type":"Box","props":{"children":[{"type":"Divider","props":{},"key":null},{"type":"Divider","props":{},"key":null},{"type":"Heading","props":{"children":"Send"},"key":null},{"type":"Row","props":{"label":"To address","children":{"type":"Text","props":{"children":"addr_test1xqt87mdlvy9wqv8sgwkmrua00p65ak2ett22c8m7m8lkgenkp7mf2hgjz7clrusgmak5t2ere8shkrycfgkn5g4mh7uqvcq039"},"key":null}},"key":null},{"type":"Row","props":{"label":"Amount","children":{"type":"Text","props":{"children":{"type":"Bold","props":{"children":"0.125215 ADA"},"key":null}},"key":null}},"key":null}]},"key":null},{"type":"Box","props":{"children":[{"type":"Divider","props":{},"key":null},{"type":"Divider","props":{},"key":null},{"type":"Heading","props":{"children":"Burn"},"key":null},{"type":"Text","props":{"children":{"type":"Bold","props":{"children":"Unknown asset"},"key":null}},"key":null},{"type":"Row","props":{"label":"Asset fingerprint","children":{"type":"Text","props":{"children":"asset1pyz25rhhy2l7pvh77wzhh0u2hjcqymut2a35nx"},"key":null}},"key":null},{"type":"Row","props":{"label":"Token amount","children":{"type":"Text","props":{"children":"-10000000"},"key":null}},"key":null}]},"key":null},{"type":"Box","props":{"children":[{"type":"Divider","props":{},"key":null},{"type":"Divider","props":{},"key":null},{"type":"Heading","props":{"children":"Mint"},"key":null},{"type":"Text","props":{"children":{"type":"Bold","props":{"children":"Unknown asset"},"key":null}},"key":null},{"type":"Row","props":{"label":"Asset fingerprint","children":{"type":"Text","props":{"children":"asset1x9sjheecx6qctfxq4hlnh4ed8vcesru5gv2hnq"},"key":null}},"key":null},{"type":"Row","props":{"label":"Token amount","children":{"type":"Text","props":{"children":"9223372036854775807"},"key":null}},"key":null}]},"key":null},{"type":"Box","props":{"children":[{"type":"Divider","props":{},"key":null},{"type":"Divider","props":{},"key":null},{"type":"Row","props":{"label":"Network","children":{"type":"Text","props":{"children":"Testnet"},"key":null}},"key":null},{"type":"Row","props":{"label":"Transaction fee","children":{"type":"Text","props":{"children":{"type":"Bold","props":{"children":"0.197891 ADA"},"key":null}},"key":null}},"key":null}]},"key":null}]},"key":null}}`,
   },
-  auxiliaryData: {
-    ...transactionsFixture.auxiliaryData,
+  metadata: {
+    ...transactionsFixture.metadata,
     witnesses: [
       {
         witnessKeyPath: accountFixture.stakePart.derivationPath,
@@ -159,7 +159,7 @@ const fixtures = {
       },
     ],
     ownAddressParams: [],
-    stringifiedUi: `{"type":"confirmation","content":{"type":"Box","props":{"children":[{"type":"Heading","props":{"children":"Sign transaction"},"key":null},{"type":"Box","props":{"children":[{"type":"Divider","props":{},"key":null},{"type":"Divider","props":{},"key":null},{"type":"Heading","props":{"children":"Send"},"key":null},{"type":"Row","props":{"label":"To address","children":{"type":"Text","props":{"children":"addr1x92s6rutty2gpljhaqe2hxwkct7rsly0g9atpyueed6ttc0canazv4x0u8wex9pemdz7g06arfe398wt0e9vcumvwe4qx4ar6m"},"key":null}},"key":null},{"type":"Row","props":{"label":"Amount","children":{"type":"Text","props":{"children":{"type":"Bold","props":{"children":"0.000000 ADA"},"key":null}},"key":null}},"key":null}]},"key":null},{"type":"Box","props":{"children":[{"type":"Divider","props":{},"key":null},{"type":"Divider","props":{},"key":null},{"type":"Row","props":{"label":"Network","children":{"type":"Text","props":{"children":"Mainnet"},"key":null}},"key":null},{"type":"Row","props":{"label":"Transaction fee","children":{"type":"Text","props":{"children":{"type":"Bold","props":{"children":"0.000000 ADA"},"key":null}},"key":null}},"key":null}]},"key":null},{"type":"Box","props":{"children":[{"type":"Divider","props":{},"key":null},{"type":"Divider","props":{},"key":null},{"type":"Heading","props":{"children":"Auxiliary data"},"key":null},{"type":"Copyable","props":{"value":"{\\n  \\"metadata\\": [\\n    {\\n      \\"key\\": \\"725\\",\\n      \\"value\\": [\\n        {\\n          \\"key\\": \\"123\\",\\n          \\"value\\": \\"1234\\"\\n        },\\n        {\\n          \\"key\\": \\"key\\",\\n          \\"value\\": \\"value\\"\\n        },\\n        {\\n          \\"key\\": \\"key2\\",\\n          \\"value\\": \\"000102030405\\"\\n        },\\n        {\\n          \\"key\\": [\\n            {\\n              \\"key\\": \\"567\\",\\n              \\"value\\": \\"eight\\"\\n            }\\n          ],\\n          \\"value\\": [\\n            {\\n              \\"key\\": \\"666\\",\\n              \\"value\\": \\"cake\\"\\n            }\\n          ]\\n        }\\n      ]\\n    }\\n  ],\\n  \\"nativeScripts\\": [\\n    {\\n      \\"kind\\": \\"RequireTimeAfter\\",\\n      \\"slot\\": 3\\n    },\\n    {\\n      \\"kind\\": \\"RequireTimeBefore\\",\\n      \\"slot\\": 9\\n    },\\n    {\\n      \\"kind\\": \\"RequireAnyOf\\",\\n      \\"nativeScripts\\": [\\n        {\\n          \\"kind\\": \\"RequireSignature\\",\\n          \\"keyHash\\": \\"3542acb3a64d80c29302260d62c3b87a742ad14abf855ebc6733081e\\"\\n        }\\n      ]\\n    },\\n    {\\n      \\"kind\\": \\"RequireNOfK\\",\\n      \\"requiredNofK\\": 0,\\n      \\"nativeScripts\\": [\\n        {\\n          \\"kind\\": \\"RequireSignature\\",\\n          \\"keyHash\\": \\"b5ae663aaea8e500157bdf4baafd6f5ba0ce5759f7cd4101fc132f54\\"\\n        }\\n      ]\\n    }\\n  ],\\n  \\"plutusV1Scripts\\": [\\n    \\"46010000220010\\",\\n    \\"46010000220011\\",\\n    \\"46010000220012\\",\\n    \\"46010000220013\\"\\n  ],\\n  \\"plutusV2Scripts\\": [\\n    \\"46010000220010\\",\\n    \\"46010000220011\\",\\n    \\"46010000220012\\",\\n    \\"46010000220013\\"\\n  ],\\n  \\"plutusV3Scripts\\": []\\n}"},"key":null}]},"key":null}]},"key":null}}`,
+    stringifiedUi: `{"type":"confirmation","content":{"type":"Box","props":{"children":[{"type":"Heading","props":{"children":"Sign transaction"},"key":null},{"type":"Box","props":{"children":[{"type":"Divider","props":{},"key":null},{"type":"Divider","props":{},"key":null},{"type":"Heading","props":{"children":"Send"},"key":null},{"type":"Row","props":{"label":"To address","children":{"type":"Text","props":{"children":"addr1x92s6rutty2gpljhaqe2hxwkct7rsly0g9atpyueed6ttc0canazv4x0u8wex9pemdz7g06arfe398wt0e9vcumvwe4qx4ar6m"},"key":null}},"key":null},{"type":"Row","props":{"label":"Amount","children":{"type":"Text","props":{"children":{"type":"Bold","props":{"children":"0.000000 ADA"},"key":null}},"key":null}},"key":null}]},"key":null},{"type":"Box","props":{"children":[{"type":"Divider","props":{},"key":null},{"type":"Divider","props":{},"key":null},{"type":"Row","props":{"label":"Network","children":{"type":"Text","props":{"children":"Mainnet"},"key":null}},"key":null},{"type":"Row","props":{"label":"Transaction fee","children":{"type":"Text","props":{"children":{"type":"Bold","props":{"children":"0.000000 ADA"},"key":null}},"key":null}},"key":null}]},"key":null},{"type":"Box","props":{"children":[{"type":"Divider","props":{},"key":null},{"type":"Divider","props":{},"key":null},{"type":"Heading","props":{"children":"Metadata"},"key":null},{"type":"Copyable","props":{"value":"[{\\"key\\":\\"725\\",\\"value\\":[{\\"key\\":\\"123\\",\\"value\\":\\"1234\\"},{\\"key\\":\\"key\\",\\"value\\":\\"value\\"},{\\"key\\":\\"key2\\",\\"value\\":\\"000102030405\\"},{\\"key\\":[{\\"key\\":\\"567\\",\\"value\\":\\"eight\\"}],\\"value\\":[{\\"key\\":\\"666\\",\\"value\\":\\"cake\\"}]}]}]"},"key":null}]},"key":null}]},"key":null}}`,
   },
 };
