+++ packages/cardano-metamask-snap/src/fixtures/transactions.ts
@@ -69,8 +69,8 @@ export const transactionsFixture = {
       'd2cb8c564b73d4b6a7fc0708244aa413c4dd68c5d206b7bce8d9de4601395135',
     networkId: 0,
   },
-  auxiliaryData: {
-    // auxiliaryData from https://github.com/input-output-hk/cardano-js-sdk/blob/cbbede75349c359688c91c963deab0a8a9c8cd24/packages/core/test/Serialization/AuxiliaryData.test.ts
+  metadata: {
+    // metadata from https://github.com/input-output-hk/cardano-js-sdk/blob/cbbede75349c359688c91c963deab0a8a9c8cd24/packages/core/test/Serialization/AuxiliaryData.test.ts
     txCborHex:
       '83a3008282582014461e17271b4a108f679eb7b6947aea29573296a5edca635d583fb40785e05d0082582094461e17271b4a108f679eb7b6947aea29573296a5edca635d583fb40785e05d00018182583931550d0f8b591480fe57e832ab99d6c2fc387c8f417ab09399cb74b5e1f8ecfa2654cfe1dd931439db45e43f5d1a73129dcb7e4acc736c766a000200a20081825820abd0f26723a5de57c10eb483b14c0aec1c365d911d46ab38684c2b9b2fa4a4915840f2b04185587ed5af88cac6778b0a8392f1cd4d51e6c3722d96db62cae9d716f2d71a22aac6bde7ec097e1357b9e2ffa70eb9ab5d757d24180c843593fb302f0901828201828200581cc4b9265645fde9536c0795adbcc5291767a0c61fd62448341d7e03868200581ce01b7ece78d656ad5848362ded335254167378c1723cd94df336a6308200581c7ed7fe51d02aede226df3912f4f347bf9598138091801119a3dc7a1fd90103a400a11902d5a4187b1904d2636b65796576616c7565646b65793246000102030405a1190237656569676874a119029a6463616b6501848204038205098202818200581c3542acb3a64d80c29302260d62c3b87a742ad14abf855ebc6733081e830300818200581cb5ae663aaea8e500157bdf4baafd6f5ba0ce5759f7cd4101fc132f540284474601000022001047460100002200114746010000220012474601000022001303844746010000220010474601000022001147460100002200124746010000220013',
     txBodyHashHex:
